rename util.js => utils.js, since it is a more common name
